+++ src/components/Icon/Icons/index.tsx
@@ -1,11 +1,15 @@
 import IcCheck from "./IcCheck";
 import IcClose from "./IcClose";
+import IcEye from "./IcEye";
+import IcEyeSlash from "./IcEyeSlash";
 import { IcPlus } from "./IcPlus";
 
 const Icons = {
 	IcCheck: <IcCheck />,
 	IcClose: <IcClose />,
 	IcPlus: <IcPlus />,
+	IcEye: <IcEye />,
+	IcEyeSlash: <IcEyeSlash />,
 };
 
 export default Icons;