+++ src/components/Switch/types.ts
@@ -1,0 +1,1 @@
+export type TSwitchSize = "small" | "regular" | "large";