+++ src/components/Textarea/index.ts
@@ -1,0 +1,3 @@
+export { Textarea } from "./Textarea.component";
+export type { ITextareaProps } from "./Textarea.component";
+import "./Textarea.scss";