+++ src/components/index.ts
@@ -1,7 +1,11 @@
+export * from "./Badge";
 export * from "./Button";
 export * from "./Checkbox";
 export * from "./Grid";
 export * from "./Icon";
 export * from "./Input";
+export * from "./Radio";
+export * from "./Switch";
+export * from "./Textarea";
 export * from "./Typography";
 import "../styles/globals.scss";